requirements.txt
streamlit
google-generativeai==0.4.1

